Lots of bricks, still no bounce

+++ src/render_state.c
@@ -51,7 +51,7 @@ void drawBall(state *s) {
 
 void drawBricks(state *s) {
     SDL_SetRenderDrawColor(s->renderer, 0x00, 0xFF, 0xFF, 0xFF);
-    for (Uint8 i = 0; i < s->brick_count; i ++) {
+    for (Uint32 i = 0; i < s->brick_count; i ++) {
         brick *b = &s->bricks[i];
         if (b->showing) {
             SDL_RenderFillRect(s->renderer, &b->rect);
